[REF] l10n_ch: Demo Data Cleanup
Up until now, we did not have any common demo data package for localizations.
This is an issue for localizations having both accounting and hr packages,
because these do not depend on each others.
It means that to have complete demo data individually, they need to define
demo company in both cases.

This leads to duplication in demo companies for a single country when both
packages are installed, which is not ideal for testing as it also means that
demo data is split between multiple companies.

l10n_ch as it is known today becomes l10n_ch_account (as it was all along
an invoicing module).
l10n_ch is instead a new module containing the common demo data.

task id # 3555251
