+++ addons/l10n_ch_account/demo/account_cash_rounding.xml
