+++ addons/l10n_ch_account/demo/demo_company.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <function model="account.chart.template" name="try_loading">
+        <value eval="[]"/>
+        <value>ch</value>
+        <value model="res.company" eval="obj().env.ref('l10n_ch.demo_company_ch')"/>
+    </function>
+</odoo>
