+++ addons/l10n_ch_account/i18n/l10n_ch.pot
