+++ addons/l10n_ch_account/i18n/zh_CN.po
