+++ addons/l10n_ch_account/models/account_journal.py
