+++ addons/l10n_ch_account/models/res_bank.py
