+++ addons/l10n_ch_account/static/src/font/LiberationSans-Bold.woff
