+++ addons/l10n_ch_account/static/src/font/liberation_sans_license.txt
