+++ addons/l10n_ch_account/static/src/img/CH-Cross_7mm.png
