+++ addons/l10n_ch_account/static/src/img/scissors_h.png
