+++ addons/l10n_ch_account/static/src/scss/report_swissqr.scss
