+++ addons/l10n_ch_account/tests/test_ch_qr_code.py
