+++ addons/l10n_ch_account/tests/test_l10n_ch_qr_print.py
