+++ addons/l10n_ch_account/tests/test_swissqr.py
