+++ addons/l10n_ch_account/views/account_invoice.xml
