+++ addons/l10n_ch_account/views/account_payment_view.xml
