+++ addons/l10n_ch_account/views/qr_invoice_wizard_view.xml
