+++ addons/l10n_ch_account/views/res_bank_view.xml
