+++ addons/l10n_ch_account/views/setup_wizard_views.xml
