+++ addons/l10n_ch_account/wizard/qr_invoice_wizard.py
