+++ addons/l10n_ch_account/wizard/setup_wizards.py
